Infer operator return types without generating code
Return-type inference ran a full LTO-IR compilation and discarded the
generated code, around 470 ms per operator. It also had to be told a target
compute capability, because generating code makes the backend resolve one from
the current device -- inference itself does not depend on the target.

Infer the type with the compile-only path instead, so inference needs neither
a target arch nor a device.
